% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{as.igraph.tsn_network}
\alias{as.igraph.tsn_network}
\title{Coerce a \code{tsn_network} Object into an \code{igraph} Object.}
\usage{
\method{as.igraph}{tsn_network}(x, mode = "undirected", ...)
}
\arguments{
\item{x}{A \code{tsn_network} object.}

\item{mode}{Character scalar, specifies how igraph should interpret the
supplied matrix. See also the \code{weighted} argument, the interpretation
depends on that too. Possible values are: \code{directed},
\code{undirected}, \code{upper}, \code{lower}, \code{max}, \code{min},
\code{plus}. See details below.}

\item{...}{Ignored.}
}
\value{
An \code{igraph} object.
}
\description{
Coerce a \code{tsn_network} Object into an \code{igraph} Object.
}
