% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{as.matrix.tsn}
\alias{as.matrix.tsn}
\title{Coerce a TSN result to a weighted adjacency matrix}
\usage{
\method{as.matrix}{tsn}(x, rownames.force = NA, ...)
}
\arguments{
\item{x}{A \code{tsn} result.}

\item{rownames.force}{Ignored.}

\item{...}{Ignored.}
}
\value{
A named numeric matrix.
}
\description{
Coerce a TSN result to a weighted adjacency matrix
}
