% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_series}
\alias{plot_series}
\title{Plot Time-Series Data with State Frequencies}
\usage{
plot_series(
  data,
  selected,
  overlay = "v",
  points = FALSE,
  ncol = NULL,
  max_series = 10,
  trend = FALSE,
  scales = c("free", "free_x", "free_y", "fixed")
)
}
\arguments{
\item{data}{[\code{tsn}, \code{ts}, \code{data.frame}, \code{numeric()}]\cr
Time-series data to be plotted.}

\item{selected}{[\code{character()}]\cr A vector of indices or names of
individual time-series to plot. If not provided (default), all
time-series are plotted up to \code{max_series} number of plots.}

\item{overlay}{[\code{logical(1)}]\cr An option for plotting the overlay
that indicates the state assigned to each time point. Can be either \code{"h"}
for a horizontal overlay or \code{"v"} for a vertical overlay (default). If
\code{NULL}, no overlay is plotted.}

\item{points}{[\code{logical(1)}]\cr Should a point be added for each
observation? Defaults to \code{FALSE}. The points are colored according to the
assigned state.}

\item{ncol}{An \code{integer} giving the number of columns to use for the facets.}

\item{max_series}{[\code{integer(1)}]\cr The maximum number  of time-series to
plot. The default is 10.}

\item{trend}{[\code{logical(1)}]\cr Should are trend line be added to the plot?
The default is \code{FALSE} for no trend line.}

\item{scales}{[\code{character(1)}]\cr Any of \code{"fixed"}, \code{"free_x"}, \code{"free_y"},
or \code{"free"} (default).}
}
\value{
A \code{ggplot} object.
}
\description{
Plot Time-Series Data with State Frequencies
}
\examples{
ts_data <- data.frame(
  id = gl(10, 100),
  time = rep(1:100,10),
  series = c(
    replicate(
      10,
      stats::arima.sim(list(order = c(2, 1, 0), ar = c(0.5, 0.2)), n = 99)
    )
  )
)

ts_data_disc <- discretize(ts_data, "id", "series", "time", n_states = 5)
plot_series(ts_data_disc)

}
