% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{print.tsn}
\alias{print.tsn}
\title{Print a TSN network}
\usage{
\method{print}{tsn}(x, ...)
}
\arguments{
\item{x}{A \code{tsn} result.}

\item{...}{Additional arguments passed to \code{print.data.frame()}.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a TSN network
}
