% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ews.R
\name{detect_warnings}
\alias{detect_warnings}
\title{Detect Early Warning Signals in a Time Series}
\usage{
detect_warnings(
  data,
  ts_col,
  time_col,
  method = "rolling",
  metrics = "all",
  window = 50,
  burnin = 30,
  detrend = "none",
  threshold = 2,
  bandwidth,
  span,
  degree,
  ...
)
}
\arguments{
\item{data}{A numeric vector or a data frame.}

\item{ts_col}{Character. The name of the value column (if \code{data} is a data frame).}

\item{time_col}{Character. The name of the time column (if \code{data} is a data frame).}

\item{method}{Character. The analysis method: \code{"rolling"} or \code{"expanding"}.}

\item{metrics}{Character vector. EWS metrics to calculate. Default: \code{"all"}.}

\item{window}{Integer (percentage). Window size for the rolling method.}

\item{burnin}{Integer. Burn-in period for the expanding method.}

\item{detrend}{TODO}

\item{threshold}{Numeric. For \code{expanding}, the Z-score threshold. For \code{rolling},
the Kendall Tau threshold for a strong signal. Defaults are \code{2} and \code{0.7}.}

\item{tail.direction}{Character. Test direction for the expanding method.}
}
\value{
An object of class \code{detect_warning} containing the EWS results,
parameters, original data, and a summary.
}
\description{
Detect Early Warning Signals in a Time Series
}
