% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsn-package.R
\docType{data}
\name{motivation}
\alias{motivation}
\title{Motivation data}
\format{
A \code{data.frame} object.
}
\source{
\url{https://osf.io/preprints/psyarxiv/qxa4n_v2}
}
\usage{
motivation
}
\description{
TODO describe
}
\keyword{datasets}
