% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsn-package.R
\docType{data}
\name{steps}
\alias{steps}
\title{Step count data}
\format{
A \code{data.frame} object.
}
\source{
\url{https://pmc.ncbi.nlm.nih.gov/articles/PMC8121346/}
}
\usage{
steps
}
\description{
TODO describe
}
\keyword{datasets}
