% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{summary.tsn}
\alias{summary.tsn}
\title{Summarize a TSN network}
\usage{
\method{summary}{tsn}(object, ...)
}
\arguments{
\item{object}{A \code{tsn} result.}

\item{...}{Ignored.}
}
\value{
A one-row data frame.
}
\description{
Summarize a TSN network
}
