% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{tsn-package}
\alias{tsn-package}
\title{tsn: Time-Series Network Construction}
\description{
\code{tsn} constructs distance, visibility, and state-transition networks from
time-series data. \code{\link[=tsn]{tsn()}} is the core network constructor; \code{\link[=vg]{vg()}},
\code{\link[=discretize]{discretize()}}, \code{\link[=trend]{trend()}}, and the \code{ts_tna()} family provide focused
workflows. Core construction uses only packages distributed with R, while
transition-network models require the suggested Nestimate package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/santikka/tsn/}
  \item Report bugs at \url{https://github.com/santikka/tsn/issues/}
}

}
\author{
\strong{Maintainer}: Santtu Tikka \email{santtuth@gmail.com}

Authors:
\itemize{
  \item Mohammed Saqr
  \item Sonsoles López-Pernas
}

}
\keyword{internal}
