% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsn-package.R
\docType{package}
\name{tsn-package}
\alias{tsn}
\alias{tsn-package}
\title{The \code{tsn} Package.}
\description{
Provides functions for time-series network analysis including
distance calculation, network construction, and visualization. Provides
multiple network construction methods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/santikka/tsn/}
  \item Report bugs at \url{https://github.com/santikka/tsn/issues/}
}

}
\author{
Mohammed Saqr, Santtu Tikka, Sonsoles López-Pernas
}
