% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsn-package.R
\docType{package}
\name{tsn-package}
\alias{tsn}
\alias{tsn-package}
\title{The \code{tsn} Package.}
\description{
Performs time-series discretization and transformation to
networks. Supports discretization of univariate time series data by binning,
kernel density estimation, gaussian mixtures, and k-means among others.
Various methods for distance-based network transformations are supported
including k-nearest neighbors and gaussian kernels.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/santikka/tsn/}
  \item Report bugs at \url{https://github.com/santikka/tsn/issues/}
}

}
\author{
Santtu Tikka, Mohammed Saqr, Sonsoles López-Pernas
}
